requirements.txt
openai
python-dotenv
click
rich
streamlit
